Move next to main bundle, old bundle to /legacy
