+++ test/legacy.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
 
-import { FormComponentProps, Formik, FormikProps } from '../src/next';
+import { Formik, InjectedFormikProps } from '../src/legacy';
 import { mount, shallow } from 'enzyme';
 
 const Yup = require('yup');
@@ -22,7 +22,7 @@ interface Values {
   name: string;
 }
 
-const Form: React.SFC<FormComponentProps<Values>> = ({
+const Form: React.SFC<InjectedFormikProps<Props, Values>> = ({
   values,
   handleSubmit,
   handleChange,
@@ -62,17 +62,18 @@ const Form: React.SFC<FormComponentProps<Values>> = ({
   );
 };
 
-const BasicForm = (
-  <Formik
-    getInitialValues={{ name: 'jared' }}
-    handleSubmit={noop}
-    component={Form}
-  />
-);
+const FormFactory = (options = {}) =>
+  Formik<Props, Values, Values>({
+    mapPropsToValues: ({ user }) => ({ ...user }),
+    handleSubmit: noop,
+    ...options,
+  })(Form);
 
-describe('Formik Next', () => {
+const BasicForm = FormFactory();
+
+describe('Formik', () => {
   it('should initialize Formik state and pass down props', () => {
-    const tree = shallow(BasicForm);
+    const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
     expect(tree.find(Form).props().isSubmitting).toBe(false);
     expect(tree.find(Form).props().touched).toEqual({});
     expect(tree.find(Form).props().values).toEqual({ name: 'jared' });
@@ -84,7 +85,7 @@ describe('Formik Next', () => {
   describe('FormikHandlers', () => {
     describe('handleChange', () => {
       it('sets values state', async () => {
-        const tree = shallow(BasicForm);
+        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
 
         // Simulate a change event in the inner Form component's input
         tree.find(Form).dive().find('input').simulate('change', {
@@ -102,7 +103,7 @@ describe('Formik Next', () => {
       });
 
       it('updates values state via `name` instead of `id` attribute when both are present', async () => {
-        const tree = shallow(BasicForm);
+        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
 
         // Simulate a change event in the inner Form component's input
         tree.find(Form).dive().find('input').simulate('change', {
@@ -122,15 +123,11 @@ describe('Formik Next', () => {
 
       it('runs validations if validateOnChange is set to true', async () => {
         const validate = jest.fn(noop);
-        const tree = shallow(
-          <Formik
-            getInitialValues={{ name: 'jared' }}
-            handleSubmit={noop}
-            component={Form}
-            validate={validate}
-            validateOnChange={true}
-          />
-        );
+        const ValidationForm = FormFactory({
+          validate,
+          validateOnChange: true,
+        });
+        const tree = shallow(<ValidationForm user={{ name: 'jared' }} />);
         tree.find(Form).dive().find('input').simulate('change', {
           persist: noop,
           target: {
@@ -141,18 +138,10 @@ describe('Formik Next', () => {
         expect(validate).toHaveBeenCalled();
       });
 
-      it('does NOT run validations if validateOnChange is set to false', async () => {
+      it('does NOT run validations by default or if validateOnChange is set to false', async () => {
         const validate = jest.fn(noop);
-
-        const tree = shallow(
-          <Formik
-            getInitialValues={{ name: 'jared' }}
-            handleSubmit={noop}
-            component={Form}
-            validate={validate}
-            validateOnChange={false}
-          />
-        );
+        const ValidationForm = FormFactory({ validate });
+        const tree = shallow(<ValidationForm user={{ name: 'jared' }} />);
         tree.find(Form).dive().find('input').simulate('change', {
           persist: noop,
           target: {
@@ -166,7 +155,7 @@ describe('Formik Next', () => {
 
     describe('handleBlur', () => {
       it('sets touched state', () => {
-        const tree = shallow(BasicForm);
+        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
 
         // Simulate a blur event in the inner Form component's input
         tree.find(Form).dive().find('input').simulate('blur', {
@@ -179,7 +168,7 @@ describe('Formik Next', () => {
       });
 
       it('updates touched state via `name` instead of `id` attribute when both are present', () => {
-        const tree = shallow(BasicForm);
+        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
 
         // Simulate a blur event in the inner Form component's input
         tree.find(Form).dive().find('input').simulate('blur', {
@@ -192,32 +181,31 @@ describe('Formik Next', () => {
         expect(tree.update().state().touched).toEqual({ name: true });
       });
 
-      it('runs validations if validateOnBlur is set to true ', async () => {
+      it('runs validations by default or if validateOnBlur is set to true ', async () => {
         const validate = jest.fn(noop);
-
-        const tree = shallow(
-          <Formik
-            getInitialValues={{ name: 'jared' }}
-            handleSubmit={noop}
-            component={Form}
-            validate={validate}
-            validateOnBlur={true}
-          />
-        );
-
+        const ValidationForm = FormFactory({ validate });
+        const ValidationForm2 = FormFactory({ validate, validateOnBlur: true });
+        const tree = shallow(<ValidationForm user={{ name: 'jared' }} />);
+        const tree2 = shallow(<ValidationForm2 user={{ name: 'jared' }} />);
         tree.find(Form).dive().find('input').simulate('blur', {
           persist: noop,
           target: {
             name: 'name',
           },
         });
-        expect(validate).toHaveBeenCalledTimes(1);
+        tree2.find(Form).dive().find('input').simulate('blur', {
+          persist: noop,
+          target: {
+            name: 'name',
+          },
+        });
+        expect(validate).toHaveBeenCalledTimes(2);
       });
     });
 
     describe('handleSubmit', () => {
       it('should call preventDefault()', () => {
-        const tree = shallow(BasicForm);
+        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
         const preventDefault = jest.fn();
         tree.find(Form).dive().find('form').simulate('submit', {
           preventDefault,
@@ -226,7 +214,7 @@ describe('Formik Next', () => {
       });
 
       it('should touch all fields', () => {
-        const tree = shallow(BasicForm);
+        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
         tree.find(Form).dive().find('form').simulate('submit', {
           preventDefault: noop,
         });
@@ -234,7 +222,7 @@ describe('Formik Next', () => {
       });
 
       it('should push submission state changes to child component', () => {
-        const tree = shallow(BasicForm);
+        const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
 
         expect(tree.find(Form).dive().find('#submitting')).toHaveLength(0);
 
@@ -245,35 +233,33 @@ describe('Formik Next', () => {
         expect(tree.find(Form).dive().find('#submitting')).toHaveLength(1);
       });
 
-      // it('should correctly map form values to payload', () => {
-      //   interface Payload {
-      //     user: { name: string };
-      //   }
-      //   const CustomPayloadForm = Formik<Props, Values, Payload>({
-      //     mapPropsToValues: ({ user }) => ({ ...user }),
-      //     mapValuesToPayload: ({ name }) => ({ user: { name } }),
-      //     handleSubmit: payload => {
-      //       expect(payload).toEqual({ user: { name: 'jared' } });
-      //       expect(payload).not.toEqual({ name: 'jared' });
-      //     },
-      //   })(Form);
-      //   const tree = shallow(<CustomPayloadForm user={{ name: 'jared' }} />);
-      //   tree.find(Form).dive().find('form').simulate('submit', {
-      //     preventDefault: noop,
-      //   });
-      // });
+      it('should correctly map form values to payload', () => {
+        interface Payload {
+          user: { name: string };
+        }
+        const CustomPayloadForm = Formik<Props, Values, Payload>({
+          mapPropsToValues: ({ user }) => ({ ...user }),
+          mapValuesToPayload: ({ name }) => ({ user: { name } }),
+          handleSubmit: payload => {
+            expect(payload).toEqual({ user: { name: 'jared' } });
+            expect(payload).not.toEqual({ name: 'jared' });
+          },
+        })(Form);
+        const tree = shallow(<CustomPayloadForm user={{ name: 'jared' }} />);
+        tree.find(Form).dive().find('form').simulate('submit', {
+          preventDefault: noop,
+        });
+      });
 
       describe('with validate (SYNC)', () => {
         it('should call validate if present', () => {
           const validate = jest.fn().mockReturnValue({});
-          const tree = shallow(
-            <Formik
-              getInitialValues={{ name: 'jared' }}
-              handleSubmit={noop}
-              component={Form}
-              validate={validate}
-            />
-          );
+          const ValidateForm = Formik<Props, Values, Values>({
+            validate,
+            mapPropsToValues: ({ user }) => ({ ...user }),
+            handleSubmit: noop,
+          })(Form);
+          const tree = shallow(<ValidateForm user={{ name: 'jared' }} />);
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -282,14 +268,12 @@ describe('Formik Next', () => {
 
         it('should submit the form if valid', () => {
           const handleSubmit = jest.fn();
-          const tree = shallow(
-            <Formik
-              getInitialValues={{ name: 'jared' }}
-              handleSubmit={handleSubmit}
-              component={Form}
-              validate={noop}
-            />
-          );
+          const ValidateForm = Formik<Props, Values, Values>({
+            validate: noop,
+            mapPropsToValues: ({ user }) => ({ ...user }),
+            handleSubmit,
+          })(Form);
+          const tree = shallow(<ValidateForm user={{ name: 'jared' }} />);
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -300,14 +284,13 @@ describe('Formik Next', () => {
           const validate = jest.fn().mockReturnValue({ name: 'Error!' });
           const handleSubmit = jest.fn();
 
-          const tree = shallow(
-            <Formik
-              getInitialValues={{ name: 'jared' }}
-              handleSubmit={handleSubmit}
-              component={Form}
-              validate={validate}
-            />
-          );
+          const ValidateForm = Formik<Props, Values, Values>({
+            validate,
+            mapPropsToValues: ({ user }) => ({ ...user }),
+            handleSubmit,
+          })(Form);
+
+          const tree = shallow(<ValidateForm user={{ name: '' }} />);
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -319,15 +302,12 @@ describe('Formik Next', () => {
       describe('with validate (ASYNC)', () => {
         it('should call validate if present', () => {
           const validate = jest.fn(() => Promise.resolve({}));
-
-          const tree = shallow(
-            <Formik
-              getInitialValues={{ name: 'jared' }}
-              handleSubmit={noop}
-              component={Form}
-              validate={validate}
-            />
-          );
+          const ValidateForm = Formik<Props, Values, Values>({
+            validate,
+            mapPropsToValues: ({ user }) => ({ ...user }),
+            handleSubmit: noop,
+          })(Form);
+          const tree = shallow(<ValidateForm user={{ name: 'jared' }} />);
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -337,14 +317,13 @@ describe('Formik Next', () => {
         it('should submit the form if valid', async () => {
           const handleSubmit = jest.fn();
 
-          const tree = shallow(
-            <Formik
-              getInitialValues={{ name: 'jared' }}
-              handleSubmit={handleSubmit}
-              component={Form}
-              validate={() => Promise.resolve({})}
-            />
-          );
+          const ValidateForm = Formik<Props, Values, Values>({
+            validate: () => Promise.resolve({}),
+            mapPropsToValues: ({ user }) => ({ ...user }),
+            handleSubmit,
+          })(Form);
+
+          const tree = shallow(<ValidateForm user={{ name: '' }} />);
           await tree.find(Form).props().submitForm();
 
           expect(handleSubmit).toHaveBeenCalled();
@@ -353,17 +332,16 @@ describe('Formik Next', () => {
         it('should not submit the form if invalid', async () => {
           const handleSubmit = jest.fn();
 
-          const tree = shallow(
-            <Formik
-              getInitialValues={{ name: 'jared' }}
-              handleSubmit={handleSubmit}
-              component={Form}
-              validate={() =>
-                sleep(25).then(() => {
-                  throw { name: 'error!' };
-                })}
-            />
-          );
+          const ValidateForm = Formik<Props, Values, Values>({
+            validate: () =>
+              sleep(25).then(() => {
+                throw { name: 'error!' };
+              }),
+            mapPropsToValues: ({ user }) => ({ ...user }),
+            handleSubmit,
+          })(Form);
+
+          const tree = shallow(<ValidateForm user={{ name: '' }} />);
           await tree.find(Form).props().submitForm();
 
           expect(handleSubmit).not.toHaveBeenCalled();
@@ -373,17 +351,8 @@ describe('Formik Next', () => {
       describe('with validationSchema (ASYNC)', () => {
         it('should run validationSchema if present', () => {
           const validate = jest.fn(() => Promise.resolve({}));
-          const tree = shallow(
-            <Formik
-              getInitialValues={{ name: 'jared' }}
-              handleSubmit={noop}
-              component={Form}
-              validate={validate}
-              validationSchema={{
-                validate,
-              }}
-            />
-          );
+          const ValidateForm = FormFactory({ validationSchema: { validate } });
+          const tree = shallow(<ValidateForm user={{ name: 'jared' }} />);
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -392,17 +361,12 @@ describe('Formik Next', () => {
 
         it('should call validationSchema if it is a function and present', () => {
           const validate = jest.fn(() => Promise.resolve({}));
-          const tree = shallow(
-            <Formik
-              getInitialValues={{ name: 'jared' }}
-              handleSubmit={noop}
-              component={Form}
-              validate={validate}
-              validationSchema={() => ({
-                validate,
-              })}
-            />
-          );
+          const ValidateForm = FormFactory({
+            validationSchema: () => ({
+              validate,
+            }),
+          });
+          const tree = shallow(<ValidateForm user={{ name: 'jared' }} />);
           tree.find(Form).dive().find('form').simulate('submit', {
             preventDefault: noop,
           });
@@ -414,117 +378,84 @@ describe('Formik Next', () => {
 
   describe('FormikActions', () => {
     it('setValues sets values', async () => {
-      const tree = shallow(BasicForm);
+      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setValues({ name: 'ian' });
       expect(tree.find(Form).dive().find('input').props().value).toEqual('ian');
     });
 
     it('setValues should run validations when validateOnChange is true', async () => {
       const validate = jest.fn().mockReturnValue({});
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          validate={validate}
-          validateOnChange={true}
-        />
-      );
+      const ValidateOnBlurForm = FormFactory({
+        validate,
+        validateOnChange: true,
+      });
+      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setValues({ name: 'ian' });
       expect(validate).toHaveBeenCalled();
     });
 
-    it('setValues should NOT run validations when validateOnChange is false', () => {
+    it('setValues should NOT run validations when validateOnChange is false or undefined', () => {
       const validate = jest.fn();
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          validate={validate}
-          validateOnChange={false}
-        />
-      );
+      const ValidateOnBlurForm = FormFactory({ validate });
+      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setValues({ name: 'ian' });
       expect(validate).not.toHaveBeenCalled();
     });
 
     it('setFieldValue sets value by key', () => {
-      const tree = shallow(BasicForm);
+      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setFieldValue('name', 'ian');
       expect(tree.find(Form).dive().find('input').props().value).toEqual('ian');
     });
 
     it('setFieldValue should run validations when validateOnChange is true', () => {
       const validate = jest.fn().mockReturnValue({});
-
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          validate={validate}
-          validateOnChange={true}
-        />
-      );
+      const ValidateOnBlurForm = FormFactory({
+        validate,
+        validateOnChange: true,
+      });
+      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setFieldValue('name', 'ian');
       expect(validate).toHaveBeenCalled();
     });
 
-    it('setFieldValue should NOT run validations when validateOnChange is false', () => {
+    it('setFieldValue should NOT run validations when validateOnChange is false or undefined', () => {
       const validate = jest.fn();
-
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          validate={validate}
-          validateOnChange={false}
-        />
-      );
+      const ValidateOnBlurForm = FormFactory({ validate });
+      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setFieldValue('name', 'ian');
       expect(validate).not.toHaveBeenCalled();
     });
 
     it('setTouched sets touched', async () => {
-      const tree = shallow(BasicForm);
+      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setTouched({ name: true });
       expect(tree.find(Form).props().touched).toEqual({ name: true });
     });
 
-    it('setTouched should NOT run validations by default', async () => {
+    it('setTouched should run validations by default, or when validateOnBlur is true', async () => {
       const validate = jest.fn().mockReturnValue({});
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          validate={validate}
-        />
-      );
-      tree.find(Form).props().setTouched({ name: true });
-      expect(validate).not.toHaveBeenCalled();
-    });
-
-    it('setTouched should run validations when validateOnBlur is true', () => {
-      const validate = jest.fn();
-
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          validate={validate}
-          validateOnBlur={true}
-        />
-      );
+      const ValidateOnBlurForm = FormFactory({
+        validate,
+      });
+      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setTouched({ name: true });
       expect(validate).toHaveBeenCalled();
     });
 
+    it('setTouched should NOT run validations when validateOnBlur is false', () => {
+      const validate = jest.fn();
+      const ValidateOnBlurForm = FormFactory({
+        validate,
+        validateOnBlur: false,
+      });
+      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
+      tree.find(Form).props().setTouched({ name: true });
+      expect(validate).not.toHaveBeenCalled();
+    });
+
     it('setFieldTouched sets touched by key', async () => {
-      const tree = shallow(BasicForm);
+      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setFieldTouched('name', true);
       expect(tree.find(Form).props().touched).toEqual({ name: true });
       expect(tree.find(Form).props().dirty).toBe(true);
@@ -535,37 +466,28 @@ describe('Formik Next', () => {
 
     it('setFieldTouched should run validations when validateOnBlur is true', async () => {
       const validate = jest.fn().mockReturnValue({});
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          validate={validate}
-          validateOnBlur={true}
-        />
-      );
+      const ValidateOnBlurForm = FormFactory({
+        validate,
+        validateOnBlur: true,
+      });
+      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setFieldTouched('name', true);
       expect(validate).toHaveBeenCalled();
     });
 
     it('setFieldTouched should NOT run validations when validateOnBlur is true', async () => {
       const validate = jest.fn().mockReturnValue({});
-
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          validate={validate}
-          validateOnBlur={false}
-        />
-      );
+      const ValidateOnBlurForm = FormFactory({
+        validate,
+        validateOnBlur: false,
+      });
+      const tree = shallow(<ValidateOnBlurForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setFieldTouched('name', true);
       expect(validate).not.toHaveBeenCalled();
     });
 
     it('setErrors sets error object', async () => {
-      const tree = shallow(BasicForm);
+      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setErrors({ name: 'Required' });
       expect(tree.find(Form).dive().find('#feedback').text()).toEqual(
         'Required'
@@ -573,7 +495,7 @@ describe('Formik Next', () => {
     });
 
     it('setFieldError sets error by key', async () => {
-      const tree = shallow(BasicForm);
+      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
       tree.find(Form).props().setFieldError('name', 'Required');
       expect(tree.find(Form).dive().find('#feedback').text()).toEqual(
         'Required'
@@ -581,7 +503,7 @@ describe('Formik Next', () => {
     });
 
     it('setStatus sets status object', async () => {
-      const tree = shallow(BasicForm);
+      const tree = shallow(<BasicForm user={{ name: 'jared' }} />);
       tree.find(Form).dive().find('#statusButton').simulate('click');
       expect(tree.find(Form).dive().find('#statusMessage')).toHaveLength(1);
     });
@@ -589,73 +511,52 @@ describe('Formik Next', () => {
 
   describe('FormikComputedProps', () => {
     it('should compute dirty as soon as any input is touched', () => {
-      const tree = shallow(BasicForm);
+      const ValidForm = FormFactory();
+      const tree = shallow(<ValidForm user={{ name: 'jared' }} />);
       expect(tree.find(Form).props().dirty).toBe(false);
       tree.setState({ touched: { name: true } });
       expect(tree.find(Form).props().dirty).toBe(true);
     });
 
     it('should compute isValid if isInitialValid is present and returns true', () => {
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          isInitialValid={props => true}
-        />
-      );
+      const InvalidForm = FormFactory({ isInitialValid: props => true });
+      const tree = shallow(<InvalidForm user={{ name: 'jared' }} />);
       expect(tree.find(Form).props().dirty).toBe(false);
       expect(tree.find(Form).props().isValid).toBe(true);
     });
 
     it('should compute isValid if isInitialValid is present and returns false', () => {
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          isInitialValid={props => false}
-        />
-      );
+      const InvalidForm = FormFactory({ isInitialValid: props => false });
+      const tree = shallow(<InvalidForm user={{ name: 'jared' }} />);
       expect(tree.find(Form).props().dirty).toBe(false);
       expect(tree.find(Form).props().isValid).toBe(false);
     });
 
     it('should compute isValid if isInitialValid boolean is present and set to true', () => {
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          isInitialValid={true}
-        />
-      );
+      const InvalidForm = FormFactory({ isInitialValid: true });
+      const tree = shallow(<InvalidForm user={{ name: 'jared' }} />);
       expect(tree.find(Form).props().dirty).toBe(false);
       expect(tree.find(Form).props().isValid).toBe(true);
     });
 
     it('should compute isValid if isInitialValid is present and set to false', () => {
-      const tree = shallow(
-        <Formik
-          getInitialValues={{ name: 'jared' }}
-          handleSubmit={noop}
-          component={Form}
-          isInitialValid={false}
-        />
-      );
+      const InvalidForm = FormFactory({ isInitialValid: false });
+      const tree = shallow(<InvalidForm user={{ name: 'jared' }} />);
       expect(tree.find(Form).props().dirty).toBe(false);
       expect(tree.find(Form).props().isValid).toBe(false);
     });
 
     it('should compute isValid if the form is dirty and there are errors', () => {
-      const tree = shallow(BasicForm);
+      const ValidForm = FormFactory();
+      const tree = shallow(<ValidForm user={{ name: 'jared' }} />);
       tree.setState({ touched: { name: true }, errors: { name: 'Required!' } });
       expect(tree.find(Form).props().dirty).toBe(true);
       expect(tree.find(Form).props().isValid).toBe(false);
     });
 
     it('should compute isValid if the form is dirty and there are not errors', () => {
-      const tree = shallow(BasicForm);
+      const ValidForm = FormFactory();
+      const tree = shallow(<ValidForm user={{ name: 'jared' }} />);
       tree.setState({ touched: { name: true } });
       expect(tree.find(Form).props().dirty).toBe(true);
       expect(tree.find(Form).props().isValid).toBe(true);
